+++ NfcReader/app/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.yatapone.nfcreader"
+        applicationId "com.yatapone.samplenfcreader"
         minSdk 26
         targetSdk 32
         versionCode 1
