+++ NfcReader/app/src/androidTest/java/com/yatapone/samplenfcreader/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.yatapone.nfcreader
+package com.yatapone.samplenfcreader
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
