+++ NfcReader/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">NfcReader</string>
+    <string name="app_name">SampleNfcReader</string>
 </resources>
