+++ NfcReader/app/src/test/java/com/yatapone/samplenfcreader/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.yatapone.nfcreader
+package com.yatapone.samplenfcreader
 
 import org.junit.Test
