+++ NfcReader/settings.gradle
@@ -13,5 +13,5 @@ dependencyResolutionManagement {
     }
 }
 
-rootProject.name = "NfcReader"
+rootProject.name = "SampleNfcReader"
 include ':app'
